AMD64 - Fix bug in pmap_prefault, adjust vn kprintfs to conform to amd64.
* The pde was being indirected without first checking to see if it was NULL.

* Fix kprintf()'s in the VN device to build without warning on amd64.

+++ sys/dev/disk/vn/vn.c
@@ -571,7 +571,7 @@ vniocattach_file(struct vn_softc *vn, struct vn_ioctl *vio, cdev_t dev,
 
 	IFOPT(vn, VN_FOLLOW)
 		kprintf("vnioctl: SET vp %p size %llx blks\n",
-		       vn->sc_vp, vn->sc_size);
+		       vn->sc_vp, (long long)vn->sc_size);
 done:
 	nlookup_done(&nd);
 	return(error);
@@ -647,7 +647,7 @@ vniocattach_swap(struct vn_softc *vn, struct vn_ioctl *vio, cdev_t dev,
 	if (error == 0) {
 		IFOPT(vn, VN_FOLLOW) {
 			kprintf("vnioctl: SET vp %p size %llx\n",
-			       vn->sc_vp, vn->sc_size);
+			       vn->sc_vp, (long long)vn->sc_size);
 		}
 	}
 	if (error)
